highlight: Move client into internal package (#54617)
(cherry picked from commit d826a1e14f9dc95c1bc41081996ae38059cd3712)
