+++ cmd/frontend/graphqlbackend/BUILD.bazel
@@ -230,7 +230,6 @@ go_library(
         "//cmd/frontend/hubspot",
         "//cmd/frontend/hubspot/hubspotutil",
         "//cmd/frontend/internal/app/updatecheck",
-        "//cmd/frontend/internal/highlight",
         "//cmd/frontend/internal/processrestart",
         "//cmd/frontend/internal/search/logs",
         "//cmd/frontend/internal/siteid",
@@ -279,6 +278,7 @@ go_library(
         "//internal/goroutine/recorder",
         "//internal/gosyntect",
         "//internal/gqlutil",
+        "//internal/highlight",
         "//internal/honey",
         "//internal/honey/search",
         "//internal/httpcli",
@@ -451,7 +451,6 @@ go_test(
         "//cmd/frontend/graphqlbackend/apitest",
         "//cmd/frontend/graphqlbackend/externallink",
         "//cmd/frontend/graphqlbackend/graphqlutil",
-        "//cmd/frontend/internal/highlight",
         "//internal/actor",
         "//internal/api",
         "//internal/auth",
@@ -476,6 +475,7 @@ go_test(
         "//internal/gitserver/gitdomain",
         "//internal/gitserver/protocol",
         "//internal/gqlutil",
+        "//internal/highlight",
         "//internal/inventory",
         "//internal/oobmigration",
         "//internal/rbac",
