+++ cmd/frontend/internal/cli/BUILD.bazel
@@ -30,7 +30,6 @@ go_library(
         "//cmd/frontend/internal/app/updatecheck",
         "//cmd/frontend/internal/bg",
         "//cmd/frontend/internal/cli/middleware",
-        "//cmd/frontend/internal/highlight",
         "//cmd/frontend/internal/httpapi",
         "//cmd/frontend/internal/httpapi/router",
         "//cmd/frontend/internal/siteid",
@@ -64,6 +63,7 @@ go_library(
         "//internal/goroutine",
         "//internal/grpc",
         "//internal/grpc/defaults",
+        "//internal/highlight",
         "//internal/httpserver",
         "//internal/instrumentation",
         "//internal/jsonc",
