--- cmd/frontend/internal/search/BUILD.bazel
@@ -1,72 +1,0 @@
-load("//dev:go_defs.bzl", "go_test")
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "search",
-    srcs = [
-        "decorate.go",
-        "event_writer.go",
-        "metadata.go",
-        "search.go",
-    ],
-    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/search",
-    visibility = ["//cmd/frontend:__subpackages__"],
-    deps = [
-        "//cmd/frontend/graphqlbackend",
-        "//cmd/frontend/internal/highlight",
-        "//cmd/frontend/internal/search/logs",
-        "//internal/api",
-        "//internal/authz",
-        "//internal/conf",
-        "//internal/database",
-        "//internal/gitserver",
-        "//internal/honey",
-        "//internal/honey/search",
-        "//internal/lazyregexp",
-        "//internal/search",
-        "//internal/search/client",
-        "//internal/search/job/jobutil",
-        "//internal/search/result",
-        "//internal/search/streaming",
-        "//internal/search/streaming/api",
-        "//internal/search/streaming/client",
-        "//internal/search/streaming/http",
-        "//internal/trace",
-        "//internal/types",
-        "//lib/errors",
-        "//lib/pointers",
-        "@com_github_inconshreveable_log15//:log15",
-        "@com_github_prometheus_client_golang//prometheus",
-        "@com_github_prometheus_client_golang//prometheus/promauto",
-        "@com_github_sourcegraph_log//:log",
-        "@io_opentelemetry_go_otel//attribute",
-    ],
-)
-
-go_test(
-    name = "search_test",
-    timeout = "short",
-    srcs = [
-        "decorate_test.go",
-        "search_test.go",
-    ],
-    embed = [":search"],
-    deps = [
-        "//internal/api",
-        "//internal/database",
-        "//internal/search",
-        "//internal/search/client",
-        "//internal/search/query",
-        "//internal/search/result",
-        "//internal/search/streaming",
-        "//internal/search/streaming/api",
-        "//internal/search/streaming/http",
-        "//internal/settings",
-        "//internal/types",
-        "//schema",
-        "@com_github_google_go_cmp//cmp",
-        "@com_github_sourcegraph_log//logtest",
-        "@com_github_stretchr_testify//require",
-        "@org_golang_x_sync//errgroup",
-    ],
-)
