+++ internal/gosyntect/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/internal/gosyntect",
     visibility = ["//:__subpackages__"],
     deps = [
+        "//internal/env",
         "//internal/httpcli",
         "//internal/trace",
         "//lib/codeintel/languages",
