+++ internal/highlight/BUILD.bazel
@@ -11,14 +11,13 @@ go_library(
         "mocks.go",
         "syntect_language_map.go",
     ],
-    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/highlight",
+    importpath = "github.com/sourcegraph/sourcegraph/internal/highlight",
     visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
         "//internal/binary",
         "//internal/conf",
         "//internal/conf/conftypes",
         "//internal/conf/deploy",
-        "//internal/env",
         "//internal/gosyntect",
         "//internal/honey",
         "//internal/observation",
