+++ internal/highlight/html_test.go
