+++ internal/highlight/mocks.go
