+++ internal/highlight/testdata/telemetry-raw.txt
