+++ internal/highlight/testdata/telemetry.scip
